+++ tests/testthat/test-chart_helper.R
@@ -1,7 +1,14 @@
-test_that("check_chart_aspect_ratio works", {
+test_that("check_chart works", {
   ggplot(mtcars, aes(x = wt, y = mpg)) +
     geom_point()
 
-  expect_silent(check_chart_aspect_ratio())
+  expect_silent(check_chart())
+  expect_silent(check_chart("normal"))
+})
 
+test_that("check_chart_aspect_ratio is soft-deprecated", {
+  ggplot(mtcars, aes(x = wt, y = mpg)) +
+    geom_point()
+
+  lifecycle::expect_deprecated(check_chart_aspect_ratio())
 })